pyproject.toml
[project]
name = "multi-trading-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.47.0",
]
